requirements.txt
spacy>=3.5.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
textblob>=0.17.1
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
python-dotenv>=1.0.0
empath>=0.9
beautifulsoup4>=4.12.0
lxml>=4.9.0
